changed rating data type to match backend

+++ backend/database/trainer.sql
@@ -23,7 +23,7 @@ CREATE TABLE trainer_profile
     user_id int NOT NULL UNIQUE,
     is_public boolean NOT NULL DEFAULT false,
     price_per_hour int NOT NULL,
-    rating int,
+    rating decimal(3,2),
     philosphy varchar(50),
     bio varchar(250),
     city varchar(30),
